+++ src/mplchart/primitives/areaplot.py
@@ -54,7 +54,7 @@ class AreaPlot(Primitive):
 
         data = series_data(result, self.item, strict=True)
 
-        data = chart.extract_df(data)
+        data = chart.reindex(data)
 
         label = self.label or get_label(self.indicator)